requirements.txt
python-evtx~=0.7.4
pandas
numpy
matplotlib
plotly~=4.14.3
lxml
pyside2